+++ vanguarde/src/components/SongList.jsx
@@ -6,9 +6,8 @@ import {MDBBtn, MDBCol, MDBContainer,
 
 import {connect} from 'react-redux';
 import {SearchOptions} from './SearchOptions';
-import {toggleSearchBar} from '../reducers/searchReducer';
-import {getSongList} from '../reducers/songReducer';
-import {handleSearchChange} from '../reducers/searchReducer';
+import {toggleSearchBar, handleSearchChange} from '../redux/search';
+import {getSongList} from '../redux/songs';
 
 class SongList extends Component {
     static propTypes = {
